// Frontend/src/app/core/services/aminety-service/models/AvailableSlotResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AvailableSlotResponseDto = {
    slotId?: string;
    slotLabel?: string | null;
    slotDate?: string;
    startTime?: string;
    endTime?: string;
    maxCapacity?: number;
    currentBookings?: number;
    availableSpots?: number;
};