// Frontend/src/app/core/services/aminety-service/models/AmenityResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AmenityResponseDto = {
    id?: string;
    name?: string | null;
    slotType?: string | null;
    status?: string | null;
    location?: string | null;
    rules?: string | null;
    imageUrl?: string | null;
};